Refactor MM modal to ts
