+++ packages/web/src/pages/sign-up-page/components/ConnectedMetaMaskModal.tsx
@@ -3,7 +3,7 @@ import { useDispatch } from 'react-redux'
 
 import { configureMetaMask } from 'common/store/pages/signon/actions'
 
-import MetaMaskModal from './MetaMaskModal'
+import { MetaMaskModal } from './MetaMaskModal'
 
 const META_MASK_SETUP_URL =
   'https://help.audius.co/help/Configuring-MetaMask-For-Use-With-Audius-2d446'
